enable adhoc SetChargingProfile calls
